Build VM disk image for bootc installation VM
The image contains a disk image for the installation VM. The
installation VM can be use in order to run privileged commands unsing
rootless podman. It already contains podman, it has a systemd service
which proxies the vsock to the local podman unix socket. It also mounts
on boot a bunch of directories which contain shared mount points with
the host in order to shared the build artifacts, configurations and
the container storage.

Signed-off-by: Alice Frosi <afrosi@redhat.com>